rename inversion inputs module
