+++ runzi/configuration/subduction_inversions.py
@@ -21,7 +21,7 @@ from runzi.automation.scaling.local_config import (
 from runzi.automation.scaling.opensha_task_factory import get_factory
 from runzi.execute import subduction_inversion_solution_task
 from runzi.util.aws import get_ecs_job_config
-from runzi.runners.inversion_inputs_v2 import SubductionInversionArgs, InversionSystemArgs
+from runzi.runners.inversion_inputs import SubductionInversionArgs, InversionSystemArgs
 
 
 INITIAL_GATEWAY_PORT = 26533  # set this to ensure that concurrent scheduled tasks won't clash
